+++ CHANGELOG.md
@@ -27,6 +27,8 @@ All Notable changes to `League\Uri` will be documented in this file
 - `Uri::toUnixPath` returns the URI path as a Unix Path or `null`
 - `Uri::toWindowsPath` returns the URI path as a Windows Path or `null`
 - `Uri::toRfc8089` return the URI in a RFC8089 formator `null`
+- `Uri::toAnchor` returns the HTML anchor string using the instance as the href attribute value
+- `Uri::toMarkdown` returns the markdown link construct using the instance as the href attribute value
 
 ### Fixed